+++ dio/test/download_test.dart
@@ -11,8 +11,17 @@ import 'utils.dart';
 void main() {
   setUp(startServer);
   tearDown(stopServer);
+
+  test('download does not change the response type', () async {
+    const savePath = 'test/download/_test.md';
+    final dio = Dio()..options.baseUrl = serverUrl.toString();
+    final options = Options(responseType: ResponseType.plain);
+    await dio.download('/download', savePath, options: options);
+    expect(options.responseType, ResponseType.plain);
+  });
+
   test('download1', () async {
-    const savePath = 'test/_download_test.md';
+    const savePath = 'test/download/_test.md';
     final dio = Dio()..options.baseUrl = serverUrl.toString();
     await dio.download('/download', savePath);
 
@@ -22,7 +31,7 @@ void main() {
   });
 
   test('download2', () async {
-    const savePath = 'test/_download_test.md';
+    const savePath = 'test/download/_test.md';
     final dio = Dio()..options.baseUrl = serverUrl.toString();
     await dio.downloadUri(
       serverUrl.replace(path: '/download'),
@@ -35,7 +44,7 @@ void main() {
   });
 
   test('download error', () async {
-    const savePath = 'test/_download_test.md';
+    const savePath = 'test/download/_test.md';
     final dio = Dio()..options.baseUrl = serverUrl.toString();
     Response response = await dio
         .download('/error', savePath)
@@ -52,23 +61,35 @@ void main() {
   });
 
   test('download timeout', () async {
-    const savePath = 'test/_download_test.md';
-    final dio = Dio(
-      BaseOptions(
-        receiveTimeout: Duration(milliseconds: 1),
-        baseUrl: serverUrl.toString(),
+    final dio = Dio();
+    final timeoutMatcher = allOf([
+      throwsA(isA<DioException>()),
+      throwsA(
+        predicate<DioException>(
+          (e) => e.type == DioExceptionType.receiveTimeout,
+        ),
       ),
+    ]);
+    await expectLater(
+      dio.downloadUri(
+        Uri.parse('$serverUrl/download').replace(
+          queryParameters: {'count': '3', 'gap': '2'},
+        ),
+        'test/download/_test.md',
+        options: Options(receiveTimeout: Duration(seconds: 1)),
+      ),
+      timeoutMatcher,
     );
-    expect(
-      dio
-          .download('/download', savePath)
-          .catchError((e) => throw (e as DioException).type),
-      throwsA(DioExceptionType.receiveTimeout),
+    // Throws nothing if it constantly gets response bytes.
+    await dio.download(
+      'https://github.com/cfug/flutter.cn/archive/refs/heads/main.zip',
+      'test/download/main.zip',
+      options: Options(receiveTimeout: Duration(seconds: 1)),
     );
   });
 
   test('download cancellation', () async {
-    const savePath = 'test/_download_test.md';
+    const savePath = 'test/download/_test.md';
     final cancelToken = CancelToken();
     Future.delayed(Duration(milliseconds: 100), () {
       cancelToken.cancel();
@@ -86,7 +107,7 @@ void main() {
   });
 
   test('delete on error', () async {
-    const savePath = 'test/_download_test.md';
+    const savePath = 'test/download/_test.md';
     final f = File(savePath)..createSync(recursive: true);
     expect(f.existsSync(), isTrue);
 
@@ -106,7 +127,7 @@ void main() {
   });
 
   test('delete on cancel', () async {
-    const savePath = 'test/_download_test.md';
+    const savePath = 'test/download/_test.md';
     final f = File(savePath)..createSync(recursive: true);
     expect(f.existsSync(), isTrue);