+++ dio/test/utils.dart
@@ -91,14 +91,24 @@ Future<void> startServer() async {
       }
 
       if (path == '/download') {
+        final count = int.parse(request.uri.queryParameters['count'] ?? '1');
+        final gap = int.parse(request.uri.queryParameters['gap'] ?? '0');
         const content = 'I am a text file';
-        response.headers.set('content-encoding', 'plain');
+        final contentBytes = utf8.encode(content);
         response
-          ..statusCode = 200
-          ..contentLength = content.length
-          ..write(content);
-
-        Future.delayed(Duration(milliseconds: 300), () => response.close());
+          ..bufferOutput = gap <= 0
+          ..contentLength = contentBytes.length * count
+          ..headers.set('content-encoding', 'plain')
+          ..statusCode = 200;
+        for (int i = 0; i < count; i++) {
+          response.add(contentBytes);
+          await response.flush();
+          if (i < count - 1 && gap > 0) {
+            await Future.delayed(Duration(seconds: gap));
+          }
+        }
+        await Future.delayed(const Duration(milliseconds: 300));
+        await response.close();
         return;
       }